Project Reorganization
Reorganized the entire project into more of MVC structure. Still have
to rework model items to integrate them with existing models.
