+++ lib/promisehelpers.js
