+++ package.json
@@ -1,16 +1,17 @@
 {
   "name": "restaurant-pos",
-  "version": "1.2.1",
+  "version": "1.2.2",
   "description": "A meal cost calculator",
   "main": "public/js/index.js",
+  "server": "controllers/server.js",
   "scripts": {
     "test": "mocha ./test/ ./test/models",
     "mkdir": "mkdir -p build",
     "build": "npm run clean && npm run mkdir && npm run build:html && npm run build:css && npm run build:js",
     "watch": "npm run watch:html & npm run watch:css & npm run watch:js",
     "clean": "rm -rf build",
-    "build:html": "npm run clean:html && cp public/index.html build/ && cp public/img/* build/ && cp public/restaurant.ico build/",
-    "watch:html": "npm run build:html && chokidar public/index.html -c 'npm run build:html'",
+    "build:html": "npm run clean:html && cp views/index.html build/ && cp public/img/* build/ && cp public/restaurant.ico build/",
+    "watch:html": "npm run build:html && chokidar views/index.html -c 'npm run build:html'",
     "clean:html": "rm -f build/index.html",
     "build:css": "npm run build:css:unminified && npm run build:css:minified",
     "watch:css": "npm run watch:css:unminified & npm run watch:css:minified",
@@ -67,7 +68,7 @@
     "babel-preset-es2015": "6.9.0",
     "chai": "3.5.0",
     "chai-http": "3.0.0",
-    "chokidar-cli": "^1.2.0",
+    "chokidar-cli": "1.2.0",
     "less": "2.7.1",
     "less-plugin-clean-css": "1.5.1",
     "mocha": "2.5.3",
