+++ test/models/test-user-model.js
@@ -5,9 +5,9 @@
     var chai = require('chai');
     var chaiHttp = require('chai-http');
     //get local files
-    var server = require('../../server.js');
-    var User = require('../../server/models/user-model.js');
-    var ph = require('../../server/promisehelpers.js');
+    var server = require('../../controllers/server.js');
+    var User = require('../../models/user-model.js');
+    var ph = require('../../lib/promisehelpers.js');
     var mongoose = require('mongoose');
     //extends objects with should for test chaining
     var should = chai.should();
